+++ src/main/java/kr/co/groupworks/calendar/controller/VacationApiController.java
@@ -5,8 +5,8 @@ import kr.co.groupworks.calendar.dto.AnnualFormDTO;
 import kr.co.groupworks.calendar.dto.HalfFormDTO;
 import kr.co.groupworks.calendar.dto.OtherFormDTO;
 import kr.co.groupworks.calendar.dto.SickFormDTO;
-import kr.co.groupworks.dto.cis.employee.SessionEmployeeDTO;
 import kr.co.groupworks.calendar.service.VacationService;
+import kr.co.groupworks.employee.dto.SessionEmployeeDTO;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -74,7 +74,7 @@ public class VacationApiController {
     // 반차 신청
     @PostMapping("/half")
     public ResponseEntity<?> vacationHalf(@Validated @RequestBody HalfFormDTO dto,
-                                          @SessionAttribute(name = "employee")SessionEmployeeDTO sessionEmployeeDTO){
+                                          @SessionAttribute(name = "employee") SessionEmployeeDTO sessionEmployeeDTO){
 
         log.info("HalfFormDTO ={}",dto);
         dto.setEmployeeId(sessionEmployeeDTO.getEmployeeId());
