+++ src/main/java/kr/co/groupworks/calendar/entity/Calendar.java
@@ -2,7 +2,7 @@ package kr.co.groupworks.calendar.entity;
 
 
 import jakarta.persistence.*;
-import kr.co.groupworks.entity.cis.Employee;
+import kr.co.groupworks.employee.entity.Employee;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
 import org.springframework.data.annotation.CreatedDate;
