+++ src/main/java/kr/co/groupworks/calendar/entity/VacationHistory.java
@@ -2,8 +2,11 @@ package kr.co.groupworks.calendar.entity;
 
 import jakarta.persistence.*;
 import kr.co.groupworks.BaseEntity;
-import kr.co.groupworks.entity.cis.Employee;
-import lombok.*;
+import kr.co.groupworks.employee.entity.Employee;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import java.time.LocalDate;
 import java.time.temporal.ChronoUnit;
