+++ src/main/java/kr/co/groupworks/calendar/service/VacationServiceImpl.java
@@ -2,14 +2,17 @@ package kr.co.groupworks.calendar.service;
 
 import jakarta.persistence.EntityNotFoundException;
 import kr.co.groupworks.calendar.dto.*;
-import kr.co.groupworks.calendar.entity.*;
-import kr.co.groupworks.entity.cis.Employee;
-import kr.co.groupworks.repository.cis.EmployeeRepository;
+import kr.co.groupworks.calendar.entity.CalendarAttachment;
+import kr.co.groupworks.calendar.entity.Vacation;
+import kr.co.groupworks.calendar.entity.VacationHistory;
+import kr.co.groupworks.calendar.entity.VacationStatus;
 import kr.co.groupworks.calendar.repository.CalendarAttachmentRepository;
 import kr.co.groupworks.calendar.repository.VacationHistoryRepository;
 import kr.co.groupworks.calendar.repository.VacationRepository;
 import kr.co.groupworks.common.mapper.CalendarAttachmentMapper;
 import kr.co.groupworks.common.mapper.VacationMapper;
+import kr.co.groupworks.employee.entity.Employee;
+import kr.co.groupworks.employee.repository.EmployeeRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -19,7 +22,9 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.*;
+import java.util.Arrays;
+import java.util.List;
+import java.util.UUID;
 
 @Service
 @Slf4j
