+++ src/main/java/kr/co/groupworks/chat/entity/ChatRoomAdmin.java
@@ -1,7 +1,7 @@
 package kr.co.groupworks.chat.entity;
 
 import jakarta.persistence.*;
-import kr.co.groupworks.entity.cis.Employee;
+import kr.co.groupworks.employee.entity.Employee;
 
 @Entity
 @Table(name = "chat_room_admin")
