+++ src/main/java/kr/co/groupworks/common/security/CustomUserDetailsService.java
@@ -1,4 +1,4 @@
-package kr.co.groupworks.security;
+package kr.co.groupworks.common.security;
 
 import kr.co.groupworks.employee.entity.Employee;
 import kr.co.groupworks.employee.repository.EmployeeRepository;
