+++ src/main/resources/application.properties
@@ -24,7 +24,7 @@ spring.datasource.username=scott
 spring.datasource.password=tiger
 
 
-spring.jpa.hibernate.ddl-auto=validate
+spring.jpa.hibernate.ddl-auto=create
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.show-sql=true
